+++ OrangeDesignSystemDemo/Pages/TabBarPage.swift
@@ -39,7 +39,7 @@ struct TabBarPage: View {
                 TabBarVariant()
                 Spacer().frame(height: 10)
             }.padding(EdgeInsets(top: 0, leading: 15, bottom: 5, trailing: 15))
-        }
+        }.background(ODSColor.primaryBackground.color)
     }
 }