+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.4.0] - Unrelease
 
+- [DemoApp] Bug #155 Typography style addition, and spacing fix
+- [DemoApp/SDK] #128 Add slider action on bar tap
+- [DemoApp] #161 Design review Sprint 11 (background colours, shadows, ...) 
 - [DemoApp/SDK] #150 Add Chip component
 - [DemoApp] #42 Typographie : Update font style list (add id)
 - [DemoApp] #124 Demo App clean-up